Load product salability data on-demand #3201. Moved plugins to isAvailable only.
